fix: change table selection model
